Carry an argument's type with its value, so an unsigned one keeps its top bit
`Arg` held a value and a kind in two fields, and every integer spelling shared one kind:
`wire_int(-1)`, `wire_uint(2^64-1)` and `wire_pointer(address)` all packed as
`{bits, Wire::integer}`. The comment on it said the two integer spellings "differ in the
cast at the call site, not here" - which is true, and is the bug: the call site is the IDL
declaration, which does know whether a `uint64` is signed, and the marshaller then had to
pick a sign for a value it was handed as bits. It picked `int64`, so a value with its top
bit set went out as a negative number.

`Arg` is a `std::variant` now, so the type travels inside the value and the ambiguity cannot
be spelled: `packed_value` writes each alternative as itself, and there is no second field
to disagree with it. `wire_null` is `std::nullptr_t` (the wire's null for a value nobody
could send, and for an out-parameter nobody passed), `wire_pointer` is the unsigned integer
an address is, and `wire_real` keeps its own alternative rather than a `memcpy` into shared
storage.

The flat path needed no such machinery, because its call site already knows the type - the
generator emits `arg_uint` exactly where the IDL says `uint64` - so it only had to stop
casting: `Json(value)`, with `arg_pointer` the same for an address. That is what makes an
unsigned round-trip now, where before `as_uint64`'s own negative guard read `2^64-1` back
as `0`, which the review's claim 9 describes exactly.

Latent rather than live: every value this harness carries - a SteamID, an address, a call
counter - is under 2^63, so the wire text is unchanged for all of them and the transcripts
and the rig behave as they did. Four checks in `test_backend` pin the top-bit case and the
sign-symmetry, and they fail against a cast. MSVC x64, clang-cl and Win32 build clean and
ctest is 6/6.

The remaining `Json(static_cast<std::int64_t>(...))` calls in `session.cpp` are the same
pattern and are left alone: they carry a SteamID, a count and seconds, none of which can
lose a bit.

+++ include/bridge/call.hpp
@@ -23,13 +23,19 @@ inline Json arg_bool(bool value) noexcept { return Json(value); }
 inline Json arg_int(std::int64_t value) noexcept { return Json(value); }
 
 inline Json arg_uint(std::uint64_t value) noexcept {
-    return Json(static_cast<std::int64_t>(value));
+    // Not through an `int64`: the top bit is part of the number, and the wire has an
+    // unsigned integer for exactly this. It used to be a cast, which made
+    // `2^64-1` and `-1` the same spelling on the wire, and `reply_uint` reads the
+    // unsigned one back.
+    return Json(value);
 }
 
 inline Json arg_real(double value) noexcept { return Json(value); }
 
 inline Json arg_pointer(const void* value) noexcept {
-    return Json(static_cast<std::int64_t>(reinterpret_cast<std::uintptr_t>(value)));
+    // An address is an unsigned integer, and this is the spelling `reply_pointer`
+    // asks for back.
+    return Json(static_cast<std::uint64_t>(reinterpret_cast<std::uintptr_t>(value)));
 }
 
 inline Json arg_cstring(const char* value) { return value != nullptr ? Json(value) : Json(); }
